+++ R/package-T4transport.R
@@ -5,8 +5,6 @@
 #' @aliases T4transport-package
 #' @importFrom Rdpack reprompt
 #' @import Rdpack
-#' @import CVXR
-#' @importFrom CVXR Variable Minimize matrix_trace Problem solve
 #' @importFrom stats rnorm median dnorm quantile cov
 #' @importFrom utils packageVersion
 #' @importFrom Rcpp evalCpp